- db update from version 19 to 20 which includes data seried conversion from int to float - adjusted map for floating data series

+++ net.tourbook.device.garmin.fit/src/net/tourbook/device/garmin/fit/FitActivityContext.java
@@ -191,7 +191,7 @@ public class FitActivityContext {
 
 	private void resetSpeedAtFirstPosition(final ArrayList<TimeData> timeDataList) {
 		if (!timeDataList.isEmpty()) {
-			timeDataList.get(0).speed = Integer.MIN_VALUE;
+			timeDataList.get(0).speed = Float.MIN_VALUE;
 		}
 	}
